refactor(tls): common types
