+++ compio-tls/src/compat/mod.rs
@@ -3,5 +3,4 @@ pub(crate) mod native;
 #[cfg(feature = "py-dynamic-openssl")]
 pub(crate) mod py_ossl;
 
-#[cfg(any(feature = "native-tls", feature = "py-dynamic-openssl"))]
-mod openssl_workaround;
+mod common;
